requirements.txt
pandas
numpy
biopython
openpxyl
scipy
matplotlib
scikit-learn==1.3.0
selenium==4.11.2
chromedriver-autoinstaller

